requirements.txt
aiogram==3.10.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
pydantic==2.5.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
aiofiles==23.2.0
python-multipart==0.0.6